removing mixins
