+++ dbt_semantic_interfaces/protocols/metric.py
@@ -33,23 +33,6 @@ class MetricInputMeasure(Protocol):
         ...
 
 
-class _MetricInputMeasureMixin:
-    """Mixin class to provide common functionality for MetricInputMeasure."""
-
-    name: str
-    alias: Optional[str]
-
-    @property
-    def measure_reference(self) -> MeasureReference:
-        """Property accessor to get the MeasureReference associated with this metric input measure."""
-        return MeasureReference(element_name=self.name)
-
-    @property
-    def post_aggregation_measure_reference(self) -> MeasureReference:
-        """Property accessor to get the MeasureReference with the aliased name, if appropriate."""
-        return MeasureReference(element_name=self.alias or self.name)
-
-
 class MetricTimeWindow(Protocol):
     """Describes the window of time the metric should be accumulated over, e.g., '1 day', '2 weeks', etc."""
 
@@ -71,17 +54,6 @@ class MetricInput(Protocol):
     def as_reference(self) -> MetricReference:
         """Property accessor to get the MetricReference associated with this metric input."""
         ...
-
-
-class _MetricInputMixin:
-    """Mixin class to provide common functionality for MetricInput."""
-
-    name: str
-
-    @property
-    def as_reference(self) -> MetricReference:
-        """Property accessor to get the MetricReference associated with this metric input."""
-        return MetricReference(element_name=self.name)
 
 
 class MetricTypeParams(Protocol):
@@ -109,23 +81,6 @@ class MetricTypeParams(Protocol):
         ...
 
 
-class _MetricTypeParamsMixin:
-    """Mixin class to provide common functionality for MetricTypeParams."""
-
-    numerator: Optional[MetricInputMeasure]
-    denominator: Optional[MetricInputMeasure]
-
-    @property
-    def numerator_measure_reference(self) -> Optional[MeasureReference]:
-        """Return the measure reference, if any, associated with the metric input measure defined as the numerator."""
-        return self.numerator.measure_reference if self.numerator else None
-
-    @property
-    def denominator_measure_reference(self) -> Optional[MeasureReference]:
-        """Return the measure reference, if any, associated with the metric input measure defined as the denominator."""
-        return self.denominator.measure_reference if self.denominator else None
-
-
 class Metric(Protocol):
     """Describes a metric."""
 
@@ -137,7 +92,7 @@ class Metric(Protocol):
 
     @property
     @abstractmethod
-    def input_measures(self) -> List[MetricInputMeasure]:
+    def input_measures(self: Metric) -> List[MetricInputMeasure]:
         """Return the complete list of input measure configurations for this metric."""
         ...
 
@@ -152,34 +107,3 @@ class Metric(Protocol):
     def input_metrics(self) -> List[MetricInput]:
         """Return the associated input metrics for this metric."""
         ...
-
-
-class _MetricMixin:
-    """Some useful default implementation details of MetricProtocol."""
-
-    name: str
-    type_params: MetricTypeParams
-
-    @property
-    def input_measures(self) -> List[MetricInputMeasure]:
-        """Return the complete list of input measure configurations for this metric."""
-        tp = self.type_params
-        res = tp.measures or []
-        if tp.measure:
-            res.append(tp.measure)
-        if tp.numerator:
-            res.append(tp.numerator)
-        if tp.denominator:
-            res.append(tp.denominator)
-
-        return res
-
-    @property
-    def measure_references(self) -> List[MeasureReference]:
-        """Return the measure references associated with all input measure configurations for this metric."""
-        return [x.measure_reference for x in self.input_measures]
-
-    @property
-    def input_metrics(self) -> List[MetricInput]:
-        """Return the associated input metrics for this metric."""
-        return self.type_params.metrics or []
